🎨 Useless utils
