+++ packages/applications/projectors/src/subscribers/tâche/tâche/tâcheRenouvellée.projector.ts
@@ -1,8 +1,9 @@
-import { TâcheEntity, TâcheRenouvelléeEvent } from '@potentiel-domain/tache';
+import { TypeTâche, TâcheEntity, TâcheRenouvelléeEvent } from '@potentiel-domain/tache';
+import { DateTime } from '@potentiel-domain/common';
+import { findProjection } from '@potentiel-infrastructure/pg-projections';
+import { Option } from '@potentiel-libraries/monads';
 
 import { upsertProjection } from '../../../infrastructure';
-
-import { récupérerTâche } from './utils/récupérerTâche';
 
 export const tâcheRenouvelléeProjector = async ({
   payload: { identifiantProjet, typeTâche, ajoutéeLe },
@@ -14,4 +15,19 @@ export const tâcheRenouvelléeProjector = async ({
     typeTâche: typeTâche,
     misÀJourLe: ajoutéeLe,
   });
+};
+
+const récupérerTâche = async (typeTâche: string, identifiantProjet: string) => {
+  const tâcheEntity = await findProjection<TâcheEntity>(`tâche|${typeTâche}#${identifiantProjet}`);
+
+  const tâcheDefaultEntity: TâcheEntity = {
+    identifiantProjet,
+    typeTâche: TypeTâche.inconnue.type,
+    misÀJourLe: DateTime.now().formatter(),
+    type: 'tâche',
+  };
+
+  return Option.match(tâcheEntity)
+    .some((value) => value)
+    .none(() => tâcheDefaultEntity);
 };
